requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyngrok==7.2.12
requests==2.31.0
python-dotenv==1.0.1